services/matching/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pinecone-client==4.1.0
openai==1.35.3
psycopg2-binary==2.9.9
aio-pika==9.4.1
pydantic==2.7.4
python-dotenv==1.0.1
